api/requirements.txt
flask==3.1.1
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.1.0
deep-translator==1.11.4
gunicorn==23.0.0
ddgs>=9.14.0
python-dateutil>=2.9.0
